Patched release 0.7 for compatibility with blender 2.8
Not tested yet, I had to strip quite a few features because they
either chaned a lot between 2.7x to 2.8 or were removed
in their entirety.

Current state: The plugin does _something_ when exporting a
scene in Blender 2.80, not sure if the output is actually
correct, though...

I regret a bit working off a release archive for this patch
and not using the head-version of the original repository.