Add Particle system random object distribution
Added a section to the ogre .scene exporter to export a particle system used to layout foliage and debris randomly on a scene.

Made a document explaining how to use it and made changes to the README to reference the new document.

Needed some images to explain the proces so I created an images directory with all the images from the README as well.